+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [2.32.0] 2018-11-07
+
+### Added
+- `scrollTableOverflow` option for `<Table />`, this is now not enabled by default
+- `preventCellOverflow` option for `<Table />`, this is now not enabled by default
+- `showPageInput` option for `<Pagination />`, enables direct jump to page number
+- `disabled` option for `<Pagination />`, allows disabling the pagination controls
+- show number of results in `<Pagination />` and option to hide it
+
+### Changed
+- `<Pagination />` shows page offset by default now, can be changed by `showElementOffsetPagination` option
+
 ## [3.0.0] 2018-10-23
 
 ### Added